Rename the completive module to prophesy
