+++ lib/prophesy/src/core/prophesy.Completion.scala
@@ -30,9 +30,28 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package soundness
+package prophesy
 
-// `Completion`/`Completions` are intentionally not re-exported here: those names already belong
-// to `exoskeleton` in the `soundness` namespace. Reach them via `completive.Completion` /
-// `completive.Completions`.
-export completive.Completable
+import anticipation.*
+import denominative.*
+import stenography.*
+import vacuous.*
+
+@unexported
+object Completion:
+  enum Kind:
+    case Term, Method, Given, Extension, Type, Module, Package
+
+// A single completion candidate at a requested position. `name` is the text to
+// insert; `kind` distinguishes methods, extensions, types, etc.; `signature` is
+// the member's type rendered as a Stenography `Syntax` (a method's full
+// parameter/result signature, a value's type, …).
+case class Completion
+  ( name:          Text,
+    kind:          Completion.Kind,
+    signature:     Syntax,
+    documentation: Optional[Text] = Unset )
+
+// The result of a completion request: `replace` is the source region the chosen
+// completion replaces (an `Offset`-mode `Span`), and `items` the candidates.
+case class Completions(replace: Span, items: List[Completion])
